First working windows build. That was painful. - CMake : Use find OpenCV. - CMake : Find boost libraries (could use WX instead...) - CMake : use gnu++11 instead of c++0x to make it work in mingw - std::exception_ptr does not seem to work in mingw, commented for now   and will use a regular std::shared pointer in a future commit - explicit int to unsigned int conversion to make the compiler happy - Still lots to do...

+++ src/aweye/wx/OptionsDialog.cpp
@@ -277,14 +277,14 @@ ConfigData OptionsDialog::getData() const
             remMinSpinCtrl->GetValue() * 60 + remSecSpinCtrl->GetValue()),
         boost::posix_time::seconds(
             chkMinSpinCtrl->GetValue() * 60 + chkSecSpinCtrl->GetValue()),
-        pauseTolSpinCtrl->GetValue(),
-        workTolSpinCtrl->GetValue(),
+        (unsigned int)pauseTolSpinCtrl->GetValue(),
+        (unsigned int)workTolSpinCtrl->GetValue(),
         soundCheckBox->GetValue(),
         popupCheckBox->GetValue(),
         trayIconCheckBox->GetValue(),
         indexSpinCtrl->GetValue(),
-        FaceSizeXSpinCtrl->GetValue(),
-        FaceSizeYSpinCtrl->GetValue(),
+        (unsigned int)FaceSizeXSpinCtrl->GetValue(),
+        (unsigned int)FaceSizeYSpinCtrl->GetValue(),
         std::string(cascadeFilePickerCtrl->GetPath().mb_str()),
         std::string(soundLocFilePickerCtrl->GetPath().mb_str()),
         boost::posix_time::seconds(
